Split DRA sound to own file
