+++ config/splat.us.dra.yaml
@@ -55,12 +55,12 @@ segments:
       - [0xD0C4, .data, 7A4D0]
       - [0xD1D4, assets, factory_blueprint, g_FactoryBlueprints]
       - [0xD4B8, data]
-      - [0x1D07C, .data, 8D3E8] # g_CdVolumeTable
-      - [0x1D19C, .data, 8D3E8] # g_VolumeTable
-      - [0x1D1C0, .data, 8D3E8] # g_DebugEnabled
-      - [0x1D1C4, .data, 8D3E8] # D_800BD1C4
-      - [0x1D1C8, .data, 8D3E8] # g_VabAddrs
-      - [0x1D1E0, .data, 8D3E8] # g_SeqInfo
+      - [0x1D07C, .data, 91EBC] # g_CdVolumeTable
+      - [0x1D19C, .data, 91EBC] # g_VolumeTable
+      - [0x1D1C0, .data, 91EBC] # g_DebugEnabled
+      - [0x1D1C4, .data, 91EBC] # D_800BD1C4
+      - [0x1D1C8, .data, 91EBC] # g_VabAddrs
+      - [0x1D1E0, .data, 91EBC] # g_SeqInfo
       - [0x1D224, assets, music_xa, g_XaMusicConfigs] # g_XaMusicConfigs
       - [0x1F554, assets, sfx, g_SfxData] # g_SfxData
       - [0x2097C, data] # D_800C097C (not sure where sfx ends)
@@ -192,6 +192,7 @@ segments:
       - [0x7E4BC, c, 7E4BC]
       - [0x86ECC, c, 86ECC]
       - [0x8D3E8, c, 8D3E8]
+      - [0x91EBC, c, 91EBC]
       - [0x92F60, c, 92F60]
       - [0x93290, c, 93290]
       - [0x93BDC, c, 93BDC]
